the return type of ttyRead() and ttyWrite() functions was changed to ssize_t to enable the returning of negatives in error cases

+++ src/drivers/headers/tty.h
@@ -13,6 +13,7 @@
 #include "drivers/kbd.h"
 #include "drivers/tty.h"
 #include "ringbuf.h"
+#include "types.h"
 
 /// Declarations
 // Handles keyboard interrupts
@@ -22,10 +23,10 @@ void ttyHandleKeyboardInterrupt(uint8_t character, uint8_t keyStatuses, uint32_t
 int ttyInit(uint8_t mode);
 
 // Writes count from buf to screen
-int ttyWrite(void* buf, size_t count);
+ssize_t ttyWrite(void* buf, size_t count);
 
 // Reads count from keyboard to buf
-int ttyRead(void* buf, size_t count);
+ssize_t ttyRead(void* buf, size_t count);
 
 // Calls device specific functions
 int ttyIoctl(size_t request, ...);
